+++ Shape_detection/test/Shape_detection/test_efficient_RANSAC_torus_parameters.cpp
@@ -72,7 +72,7 @@ bool test_torus_parameters() {
     if (shapes.size() != 1)
       continue;
 
-    boost::shared_ptr<Torus> torus =
+    std::shared_ptr<Torus> torus =
       boost::dynamic_pointer_cast<Torus>((*shapes.first));
 
     // Check: shape detected is a torus.