+++ Straight_skeleton_2/examples/Straight_skeleton_2/draw_straight_skeleton_2.cpp
@@ -4,8 +4,6 @@
 #include <CGAL/create_straight_skeleton_2.h>
 #include <CGAL/draw_straight_skeleton_2.h>
 
-#include<boost/shared_ptr.hpp>
-
 #include <cassert>
 
 typedef CGAL::Exact_predicates_inexact_constructions_kernel K ;
@@ -14,7 +12,7 @@ typedef K::Point_2                   Point ;
 typedef CGAL::Polygon_2<K>           Polygon_2 ;
 typedef CGAL::Straight_skeleton_2<K> Ss ;
 
-typedef boost::shared_ptr<Ss> SsPtr ;
+typedef std::shared_ptr<Ss> SsPtr ;
 
 int main()
 {